requirements.txt
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
graphql-core==3.2.3
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
msgpack==1.0.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pydantic==1.10.2
Pygments==2.13.0
pymongo==4.3.3
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
strawberry-graphql==0.133.5
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.18.3
